Pull NL differ into a separate module
