+++ pyomo/repn/tests/ampl/test_suffixes.py
@@ -24,7 +24,7 @@ from pyomo.environ import (
     ConcreteModel, Suffix, Var, Objective, Constraint, SOSConstraint,
     sum_product,
 )
-from .test_ampl_nl import load_and_compare_nl_baseline, compare_nl_baseline
+from .nl_diff import load_and_compare_nl_baseline, compare_nl_baseline
 
 currdir = this_file_dir()
